test: remove net461; exercise API
This adds a test to exercise the `NotImplementedException` behavior of
the stubbed API. The dependency management has some thorns, so we have
removed net461 as a target framework for the time being.

We also need to forward the StackExchange.Redis types externally so
they are visible to someone who has not also installed the former
package.