feat: implement basic variant of get and set (#12)
Implements one variant each of StringGet and StringSet, adds integration tests against Redis- and Momento-backed clients, and adds CI that exercises:

a client backed by Momento on .NET 6.0
a client backed by Redis on .NET 6.0
a client backed by Momento on .NET Framework
Note we do not test a client backed by Redis on .NET Framework. Because .NET Framework runs on a windows image, and windows does not support container services, we cannot run a Redis container for the .NET Framework runs. If this becomes an issue we can look in to running it as a background process.

A future PR will address error handling in more detail.